+++ CHANGELOG.md
@@ -11,6 +11,14 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.9.0 (2020-03-28)
+
+#### :rocket: New Feature
+
+* Added `pathModifier` to `core/object/watch`
+* Added `eventFilter` to `core/object/watch`
+* Added `root` to `core/object/watch/interface/WatchHandlerParentParams`
+
 ## v3.8.0 (2020-03-25)
 
 #### :rocket: New Feature